pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["admin <admin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
drf-yasg = "^1.21.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
rest-framework-simplejwt = "^0.0.2"
pyjwt = "^2.8.0"
openai = "0.28"
django-environ = "^0.11.2"
pillow = "^10.2.0"
django = "4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
